refactor : 사용하지 않는 상수 삭제, 랜덤 숫자 생성 클래스 유틸 파라미터 변경에 따른 테스트 수정
사용하지 않는 상수 삭제, 랜덤 숫자 생성 클래스 유틸 파라미터 변경에 따른 테스트 수정
