+++ src/test/java/baseball/utils/RandomNumberUtilsTest.java
@@ -27,7 +27,7 @@ class RandomNumberUtilsTest {
     @Test
     @DisplayName("Randoms.pickNumberInRange 를 사용한 computer 숫자 3 자리")
     void createNumber01() {
-        computer = RandomNumberUtils.pickNumberInRange(start, end);
+        computer = RandomNumberUtils.pickNumberInRange(start, end, count);
 
         // 생성된 숫자의 크기 == count
         assertThat(computer.size()).isEqualTo(count);
@@ -44,7 +44,7 @@ class RandomNumberUtilsTest {
     @Test
     @DisplayName("Randoms.pickNumberInList 를 사용한 computer 숫자 3 자리")
     void createNumber02() {
-        computer = RandomNumberUtils.pickNumberInList(List.of(new Integer[]{1, 2, 3, 4, 5, 6, 7, 8, 9}));
+        computer = RandomNumberUtils.pickNumberInList(List.of(new Integer[]{1, 2, 3, 4, 5, 6, 7, 8, 9}), count);
 
         // 생성된 숫자의 크기 == count
         assertThat(computer.size()).isEqualTo(count);
